feat: updated v1 API for changes to Experiments and Jobs
This features modifies how jobs are handled in the REST API. Since Jobs
can only be created as part of an experiment, the POST has been moved to
/experiments/{id}/jobs. The GETs and DELETE associated with a specific
Job have been replicated under /experiments/{id}/jobs because Job
operations will often be performed within the context of an Experiment.
